fix strict aliasing violations
