+++ include/system/hot.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <unwind-arm-common.h>
+
 struct hot_table {
 	char const* compile_timestamp;
 	char const* compile_directory;
@@ -12,6 +14,15 @@ struct hot_table {
 #include "system/user_functions/list.h"
 #undef FUNC
 	} functions;
+};
+
+// exidx is the table that tells the unwinder how to unwind a stack frame
+// for a PC. Under hot/cold, there's two tables and the unwinder was kind
+// enough to let us implement a function to give it a table for a PC so
+// support for hot/cold is as easy as it gets
+struct __EIT_entry {
+	_uw fnoffset;
+	_uw content;
 };
 
 extern struct hot_table* const HOT_TABLE;
